migracion

+++ database/migrations/2020_02_11_035225_alter_user_table.php
@@ -14,7 +14,7 @@ class AlterUserTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('role_id')->unsigned()->index('users_role_id')->nullable(true);
+            $table->unsignedBigInteger('role_id')->unsigned()->index('users_role_id')->nullable(true);
             $table->foreign('role_id')->references('id')->on('roles');
         });
     }
